LayoutTests/imported/w3c/web-platform-tests/service-workers/stub-4.7.4.1-fetch-event-section-expected.txt

CONSOLE MESSAGE: line 871: Invalid object member name Context

Harness Error (FAIL), message = Invalid object member name Context

FAIL FetchEvent interface: existence and properties of interface object assert_own_property: self does not have own property "FetchEvent" expected property "FetchEvent" missing
FAIL FetchEvent interface object length assert_own_property: self does not have own property "FetchEvent" expected property "FetchEvent" missing
FAIL FetchEvent interface object name assert_own_property: self does not have own property "FetchEvent" expected property "FetchEvent" missing
FAIL FetchEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "FetchEvent" expected property "FetchEvent" missing
FAIL FetchEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "FetchEvent" expected property "FetchEvent" missing
FAIL FetchEvent interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "FetchEvent" expected property "FetchEvent" missing
FAIL FetchEvent interface: attribute request assert_own_property: self does not have own property "FetchEvent" expected property "FetchEvent" missing
FAIL FetchEvent interface: attribute client assert_own_property: self does not have own property "FetchEvent" expected property "FetchEvent" missing
FAIL FetchEvent interface: attribute context assert_own_property: self does not have own property "FetchEvent" expected property "FetchEvent" missing
FAIL FetchEvent interface: attribute isReload assert_own_property: self does not have own property "FetchEvent" expected property "FetchEvent" missing
FAIL FetchEvent interface: operation respondWith([object Object]) assert_own_property: self does not have own property "FetchEvent" expected property "FetchEvent" missing
FAIL FetchEvent interface: operation forwardTo(ScalarValueString) assert_own_property: self does not have own property "FetchEvent" expected property "FetchEvent" missing
FAIL FetchEvent interface: operation default() assert_own_property: self does not have own property "FetchEvent" expected property "FetchEvent" missing
FAIL FetchEvent must be primary interface of throw new Error ('No object defined for the FetchEvent interface') assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the FetchEvent interface"
FAIL Stringification of throw new Error ('No object defined for the FetchEvent interface') assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the FetchEvent interface"
FAIL FetchEvent interface: throw new Error ('No object defined for the FetchEvent interface') must inherit property "request" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the FetchEvent interface"
FAIL FetchEvent interface: throw new Error ('No object defined for the FetchEvent interface') must inherit property "client" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the FetchEvent interface"
FAIL FetchEvent interface: throw new Error ('No object defined for the FetchEvent interface') must inherit property "context" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the FetchEvent interface"
FAIL FetchEvent interface: throw new Error ('No object defined for the FetchEvent interface') must inherit property "isReload" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the FetchEvent interface"
FAIL FetchEvent interface: throw new Error ('No object defined for the FetchEvent interface') must inherit property "respondWith([object Object])" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the FetchEvent interface"
FAIL FetchEvent interface: calling respondWith([object Object]) on throw new Error ('No object defined for the FetchEvent interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the FetchEvent interface"
FAIL FetchEvent interface: throw new Error ('No object defined for the FetchEvent interface') must inherit property "forwardTo(ScalarValueString)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the FetchEvent interface"
FAIL FetchEvent interface: calling forwardTo(ScalarValueString) on throw new Error ('No object defined for the FetchEvent interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the FetchEvent interface"
FAIL FetchEvent interface: throw new Error ('No object defined for the FetchEvent interface') must inherit property "default()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the FetchEvent interface"
